Separate all commands
