+++ ikdsplit/runner.py
@@ -1,22 +1,16 @@
 """Run recursively."""
 
-import argparse
 import pathlib
+import typing
 
-from ikdsplit.filler import fill
 from ikdsplit.io import make_default_config, parse_config
-from ikdsplit.regressor import regress
-from ikdsplit.sorter import sort_all
 from ikdsplit.spacegroup import get_subgroups
-from ikdsplit.splitter import (
-    add_arguments,
-    check_criteria,
-    get_default_criteria,
-)
+from ikdsplit.splitter import check_criteria, get_default_criteria
 from ikdsplit.utils import cd, print_group
 
 
-def recur_run(
+def run_recursive(
+    fun: typing.Callable,
     config: dict,
     supergroup: int,
     group: int,
@@ -32,22 +26,17 @@ def recur_run(
         if not check_criteria(group, criteria):
             return
 
-        fill()
-
-        regress()
-
-        d = config["sort"]
-        sort_all(d["reference"])
+        fun()
 
         if 0 < criteria["max_level"] <= level:
             return
 
         subgroups = get_subgroups(group)
         for subgroup in subgroups:
-            recur_run(config, group, subgroup, level + 1, criteria)
+            run_recursive(fun, config, group, subgroup, level + 1, criteria)
 
 
-def start(criteria: dict[str, int]) -> None:
+def start_recursive(fun: typing.Callable, criteria: dict[str, int]) -> None:
     """Start calculations."""
     config = make_default_config()
     config.update(parse_config())
@@ -55,28 +44,5 @@ def start(criteria: dict[str, int]) -> None:
     criteria = get_default_criteria() | criteria
 
     print("run ...")
-    recur_run(config, None, config["space_group_number"], 0, criteria)
+    run_recursive(fun, config, None, config["space_group_number"], 0, criteria)
     print()
-
-
-def run(args: argparse.Namespace) -> None:
-    """Run."""
-    criteria = {
-        "max_level": args.level,
-        "min_order": args.order,
-        "max_configurations": args.configurations,
-    }
-    start(criteria)
-
-
-def main() -> None:
-    """Run as a script."""
-    formatter_class = argparse.ArgumentDefaultsHelpFormatter
-    parser = argparse.ArgumentParser(formatter_class=formatter_class)
-    add_arguments(parser)
-    args = parser.parse_args()
-    run(args)
-
-
-if __name__ == "__main__":
-    main()
